refactor: move ErrorBoundary and ErrorView
