+++ src/__tests__/__helpers__/renderWithProviders.tsx
@@ -21,7 +21,7 @@ import { OverlayProvider } from "react-aria";
 import { MemoryRouterProviderProps } from "next-router-mock/dist/MemoryRouterProvider/MemoryRouterProvider";
 
 import "../../browser-lib/oak-globals/oakGlobals";
-import ErrorBoundary from "../../components/ErrorBoundary";
+import ErrorBoundary from "../../components/AppComponents/ErrorBoundary";
 import { MenuProvider } from "../../context/Menu";
 import { ToastProvider } from "../../context/Toast";
 import { CookieConsentContext } from "../../browser-lib/cookie-consent/CookieConsentProvider";
