+++ src/components/AppComponents/ErrorView/ErrorView.tsx
@@ -1,12 +1,11 @@
 import { FC } from "react";
 import styled from "styled-components";
 
-import { DEFAULT_SEO_PROPS } from "../../browser-lib/seo/Seo";
+import { DEFAULT_SEO_PROPS } from "@/browser-lib/seo/Seo";
 import {
   HeaderVariant,
   FooterVariant,
-} from "../SharedComponents/Layout/Layout";
-
+} from "@/components/SharedComponents/Layout/Layout";
 import Layout from "@/components/SharedComponents/Layout";
 import { P, Heading } from "@/components/SharedComponents/Typography";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
